requirements.txt
openai
streamlit
python-dotenv
faiss-cpu
huggingface_hub
pypdf2
langchain-community
tiktoken
